src/requirements.txt
numpy>=2.0.0
tqdm>=4.60.0
python-Levenshtein>=0.20.0
ijson>=3.2.0
